pyproject.toml
[project]
name = "breedbase-client"
version = "0.1.0"
description = "BrAPI MCP server for Breedbase databases"
requires-python = ">=3.11"
dependencies = [
    "pandas",
    "requests",
    "pydantic>=2.0",
    "authlib",
    "python-dotenv",
    "mcp[cli]>=1.0.0",
    "groq>=0.33.0",
    "pytest>=9.0.1",
    "fastmcp==2.13.0.2",
    "inline-snapshot>=0.31.1",
    "pytest-asyncio>=1.3.0",
    "starlette>=0.49.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 88
indent-width = 2
unsafe-fixes = false

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
skip-magic-trailing-comma = false
